Extend language_select
Extend language_select.h to support all
locale variants supported by Java 20.

In TcMenu, it is possible to add a locale in the form lang_country.
Based on that, TcMenu generates flag like TC_LOCALE_FR_FR.
However, it won't match the definition of TC_LOCALE_FR and will
fallback to TC_LOCALE_EN despite having a better translation
available.

We're defining another variants so that the proper
locale is selected in all circumstances. Because we're
using preprocessor directives, there is no impact on the
app performance.